# frontend/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project Desert API",
    "description": "API for managing ascetical practices and spiritual growth",
    "version": "2.0.0"
  },
  "paths": {
...
    "/packages/": {
      "post": {
        "tags": ["packages"],
        "summary": "Create Package",
        "description": "Create a new asceticism package (admin only).",
        "operationId": "create_package_packages__post",
        "parameters": [
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-User-Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PackageCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PackageResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/packages/admin/all": {
      "get": {
        "tags": ["packages"],
        "summary": "Get All Packages Admin",
        "description": "Get all packages including unpublished ones (admin only).",
        "operationId": "get_all_packages_admin_packages_admin_all_get",
        "parameters": [
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/PackageResponse" },
                  "title": "Response Get All Packages Admin Packages Admin All Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/packages/{package_id}": {
      "put": {
        "tags": ["packages"],
        "summary": "Update Package",
        "description": "Update a package (admin only).",
        "operationId": "update_package_packages__package_id__put",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Package Id" }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-User-Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PackageUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PackageResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["packages"],
        "summary": "Delete Package",
        "description": "Delete a package (admin only).",
        "operationId": "delete_package_packages__package_id__delete",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Package Id" }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["packages"],
        "summary": "Get Package Details",
        "description": "Get details of a specific published package.",
        "operationId": "get_package_details_packages__package_id__get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Package Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PackageResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/packages/{package_id}/publish": {
      "post": {
        "tags": ["packages"],
        "summary": "Publish Package",
        "description": "Publish or unpublish a package (admin only).",
        "operationId": "publish_package_packages__package_id__publish_post",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Package Id" }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/packages/{package_id}/add-to-account": {
      "post": {
        "tags": ["packages"],
        "summary": "Add Package To Account",
        "description": "Add all asceticisms from a package to the user's account.",
        "operationId": "add_package_to_account_packages__package_id__add_to_account_post",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Package Id" }
          },
          {
            "name": "x-user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AsceticismInfo": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "title": { "type": "string", "title": "Title" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "category": { "type": "string", "title": "Category" },
          "icon": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Icon"
          },
          "type": { "type": "string", "title": "Type" }
        },
        "type": "object",
        "required": ["id", "title", "description", "category", "icon", "type"],
        "title": "AsceticismInfo",
        "description": "Asceticism information for package items."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PackageCreate": {
        "properties": {
          "title": { "type": "string", "title": "Title" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "custom_metadata": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Custom Metadata"
          },
          "items": {
            "items": { "$ref": "#/components/schemas/PackageItemInput" },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": ["title", "items"],
        "title": "PackageCreate",
        "description": "Request to create a package."
      },
      "PackageItemInput": {
        "properties": {
          "asceticismId": { "type": "integer", "title": "Asceticismid" },
          "order": { "type": "integer", "title": "Order", "default": 0 },
          "notes": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": ["asceticismId"],
        "title": "PackageItemInput",
        "description": "Package item input."
      },
      "PackageItemResponse": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "asceticismId": { "type": "integer", "title": "Asceticismid" },
          "order": { "type": "integer", "title": "Order" },
          "notes": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Notes"
          },
          "asceticism": { "$ref": "#/components/schemas/AsceticismInfo" }
        },
        "type": "object",
        "required": ["id", "asceticismId", "order", "notes", "asceticism"],
        "title": "PackageItemResponse",
        "description": "Package item response."
      },
      "PackageResponse": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "title": { "type": "string", "title": "Title" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "creatorId": { "type": "integer", "title": "Creatorid" },
          "isPublished": { "type": "boolean", "title": "Ispublished" },
          "custom_metadata": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Custom Metadata"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "items": {
            "items": { "$ref": "#/components/schemas/PackageItemResponse" },
            "type": "array",
            "title": "Items"
          },
          "itemCount": { "type": "integer", "title": "Itemcount" }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "creatorId",
          "isPublished",
          "custom_metadata",
          "createdAt",
          "updatedAt",
          "items",
          "itemCount"
        ],
        "title": "PackageResponse",
        "description": "Package response."
      },
      "PackageUpdate": {
        "properties": {
          "title": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Title"
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "custom_metadata": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Custom Metadata"
          },
          "items": {
            "anyOf": [
              {
                "items": { "$ref": "#/components/schemas/PackageItemInput" },
                "type": "array"
              },
              { "type": "null" }
            ],
            "title": "Items"
          }
        },
        "type": "object",
        "title": "PackageUpdate",
        "description": "Request to update a package."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}